스크립트 파트 : RDS 연동과 그에따른 수정
RDS에 접속은 간단하게 MySqlDB.cs의 값들을 수정하는 걸로 완료 되었으나
그에따라 DB의 속성 이름들과 자료형이 달라 수정이 있었다.
MySqlDB.cs : RDS db 접근을 위하여 값을 수정하였으나 민감한 값이라 실제값은 올리지 않을 것이다.
LobbyScript.cs : 이해를 위한 주석이 추가되었다.
GameHighScoreMysql.cs : 자료형과 컬럼 명이 달라저 수정하였다.
GameCharacterRankInfo.cs : 마찬가지로 자료형이 달라저 수정하였다.
GameCharacterMysql.cs : 테이블명과 자료형이 달라져서 수정하였다.
GameCharacterInfo.cs : 자료형 수정
UserInfo.cs : 자료형 수정
Accentication.cs : 자료형 수정

앞으로 할것
유저의 비번 값은 실제로 BCrypt로 암호화되어 저장되기 때문에
로그인시 암호화된 비번 값과 로그인 비번 값을 비교하는 문장을 작성할 것이다.

+++ SimpleGame/Assets/Script/Accentication/Accentication.cs
@@ -5,6 +5,8 @@ using System.Data;
 using MySql.Data.MySqlClient;
 using System;
 
+
+
 /// <summary>
 /// 회원 관련 기능 클래스
 /// </summary>
@@ -14,8 +16,9 @@ public class Accentication : MySqlDB
     /// <summary>
     /// 로그인 기능
     /// </summary>
-    public int Login(string email , string password)
+    public long Login(string email , string password)
     {
+
         connect();
         string sqlQuery = String.Format("SELECT * FROM users " +
             "where email = '{0}' AND password = '{1}'", email,password);
@@ -25,8 +28,7 @@ public class Accentication : MySqlDB
             MySqlDataReader mySqlDataReader = cmd.ExecuteReader();
             while (mySqlDataReader.Read())
             {
-
-                UserInfo.Id = (int)mySqlDataReader["id"];
+                UserInfo.Id = (long)mySqlDataReader["id"];
                 UserInfo.Email = (string)mySqlDataReader["email"];
                 UserInfo.Password = (string)mySqlDataReader["password"];
                 UserInfo.Nickname = (string)mySqlDataReader["nickname"];
