+++ SimpleGame/Assets/Script/Accentication/UserInfo.cs
@@ -2,7 +2,7 @@
 public static class UserInfo
 {
     // get,set 자동 구현
-    public static int Id { get; set; }
+    public static long Id { get; set; }
     public static string Email { get; set; }
     public static string Password { get; set; }
     public static string Nickname { get; set; }
