+++ SimpleGame/Assets/Script/GameCharacter/GameCharacterInfo.cs
@@ -4,8 +4,8 @@
 public static class GameCharacterInfo 
 {
     // get,set 자동 구현
-    public static int Id { get; set; }
+    public static long Id { get; set; }
     public static string Email { get; set; }
     public static string Nickname { get; set; }
-    public static int HighScore { get; set; }
+    public static long HighScore { get; set; }
 }
