+++ SimpleGame/Assets/Script/GameCharacter/GameCharacterRankInfo.cs
@@ -3,9 +3,9 @@ using System;
 
 public class GameCharacterRankInfo 
 {
-    public int Rank { get; set; }
+    public long Rank { get; set; }
     public string Email { get; set; }
     public string Nickname { get; set; }
-    public int HighScore { get; set; }
+    public long HighScore { get; set; }
     public DateTime LastedTime { get; set; }
 }
